pyproject.toml
[tool.poetry]
name = "openmesh-python"
version = "0.1.0"
description = "A python wrapper for Openmesh's APIs"
authors = ["mannyuncharted <omoebun52@gmail.com>"]
license = "MIT LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
websockets = "^12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
